Add HTML/Writer class
I really didn't want to do this, but I'm pretty happy with how it turned out, even though there's still a lot of finishing work to do.

Getting proper HTML out of PHP's built-ins is just an enormous pain in the ass -- maybe not even doable. Checking output from DOMDocument functions with phpunit's AssertXMLStringEqualsXMLString() is downright impossible. At least, I couldn't figure out how to do it.

The core of the problem is this: these functions are not HTML-specific. They are somewhat generic to XML. In HTML, you have some tags, like '<br>', which are self-closing tags (or, "void tags", see e.g. https://stackoverflow.com/questions/3558119/are-non-void-self-closing-tags-valid-in-html5). You also have some tags which can be empty but are NOT self-closing, like '<head></head>'.

The built-in functions don't seem to be able to handle both cases at the same time. If there's some magic incantation that makes them behave properly, I can't figure it out.

I also wanted an HTML sanitizer that can safely handle unsafe input as well as the ability to reformat html output in a consistent way without relying on tidy as an external dependency.

The HTML Writer is a good start, but needs some work still. Turning on the ENCODE_ENTITIES flag is causing a lot of mischief currently -- it seems I'm going to need to load the list of supported entities and then ignore a whole lot of them (sigh, TODO).

Many more tests need to be added for the new sanitizing and reformatting functions.